requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==3.2
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
urllib3==1.26.6
